+++ src/main/java/com/example/demo/src/user/UserController.java
@@ -103,7 +103,7 @@ public class UserController {
             return new BaseResponse<>(POST_USERS_EMPTY_PASSWORD);
         }
         try {
-            int userIdxFindByJwt = jwtService.getUserIdx();
+            Long userIdxFindByJwt = jwtService.getUserIdx();
             GetUserRes getUserRes = userProvider.getUser(userIdxFindByJwt);
             if (getUserRes.getStatus() == "N"){
                 return new BaseResponse<>(POST_USERS_DELETED_USER);
@@ -161,13 +161,13 @@ public class UserController {
     // Path-variable
     @ResponseBody
     @GetMapping("/{userIdx}") // (GET) 127.0.0.1:9000/app/users/:userIdx
-    public BaseResponse<GetUserRes> getUser(@PathVariable("userIdx") int userIdx) {
+    public BaseResponse<GetUserRes> getUser(@PathVariable("userIdx") Long userIdx) {
         // @PathVariable RESTful(URL)에서 명시된 파라미터({})를 받는 어노테이션, 이 경우 userId값을 받아옴.
         //  null값 or 공백값이 들어가는 경우는 적용하지 말 것
         //  .(dot)이 포함된 경우, .을 포함한 그 뒤가 잘려서 들어감
         // Get Users
         try {
-            int userIdxFindByJwt = jwtService.getUserIdx();
+            Long userIdxFindByJwt = jwtService.getUserIdx();
             if(userIdxFindByJwt != userIdx){
                 return new BaseResponse<>(INVALID_USER_JWT);
             }
@@ -185,11 +185,11 @@ public class UserController {
      */
     @ResponseBody
     @PatchMapping("/change-nickname/{userIdx}")
-    public BaseResponse<String> modifyUserName(@PathVariable("userIdx") int userIdx, @RequestBody User user) {
+    public BaseResponse<String> modifyUserName(@PathVariable("userIdx") Long userIdx, @RequestBody User user) {
         try {
 //  *********** 해당 부분은 7주차 - JWT 수업 후 주석해체 해주세요!  ****************
             //jwt에서 idx 추출.
-            int userIdxByJwt = jwtService.getUserIdx();
+            Long userIdxByJwt = jwtService.getUserIdx();
             //userIdx와 접근한 유저가 같은지 확인
             if(userIdx != userIdxByJwt){
                 return new BaseResponse<>(INVALID_USER_JWT);
